moved authentication related stuff to auth.go
